Added B2MML-JSON
The B2MML-JSON schema (see json-schema.org for a full definition of JSON schemas) of B2MML and BatchML was generated using XMLSpy-2021 from the AllSchemas.xsd file.  There has been no checking of this file. There has been no testing of this file. It is intended for distribution to aid in checking the use of JSON in place of XML for data exchanges.  If any errors or problems are discovered in the use or the format, then issues should be raised on the GitHub site.